Use nose to run tests from tox.ini
python setup.py test has a large downside: it creates .egg-info
directories and writes install_requires in there, and then reads it back
in order to install the dependencies.  This is a terrible idea!  It
creates a race condition between multiple Pythons running setup.py in
parallel that often breaks 'detox' by having it ask for Pygments < 2.0a1
in the Python 2.7 virtualenv.

I haven't seen that error with nose yet.