Protect against DNS rebinding attacks
Fixes #51.

The --allowed-hosts command-line argument is underdocumented (I don't
mention that the values should be space or comma-separated, nor that
glob-style wildcards * and ? are allowed).

I was unable to convince restview to listen on IPv6, so I don't know
whether ::1 or [::1] or ip6-localhost should be in the default
allowed_hosts list.