Fix tests after recent readme_renderer update
I'm not sure if it's readme_renderer or Bleach that changed behavior,
but the `clean()` helper no longer returns None in this specific
circumstance.

I want to keep the workaround in restview alive, in case there are other
circumstances or users happen to have an older readme_renderer/Bleach
version.